Modified the curl URL examples
The original URL used to demonstrate curl did not contain the application name (i.e. "jboss-contacts-angularjs").  I thought this would be confusing for students as the included url will result in a 404.  I don't think many students will change the name of their deployed .war so I figured it was safe to include the actual war name, opposed to INSERT_WAR_NAME as this might confuse them.